Restructure harmonsed integration test. Move existing integration tests to unit test folder
